+++ modules/project-templates/controller.js
