+++ modules/project-templates/init.js
